feat(replica): make NVMe-oF replica timeouts tunable via env vars
Convert hard-coded replicaCtrlrLossTimeoutSec / replicaReconnectDelaySec /
replicaFastIOFailTimeoutSec / replicaTransportAckTimeout / replicaKeepAliveTimeoutMs
from consts to vars, overridable at IM startup via LONGHORN_V2_REPLICA_* env vars.
longhorn-manager's instance_manager_controller populates those env vars from the
new data-engine-replica-* Setting CRs so operators can tune them without a code
change.

Defaults are unchanged. Only types.go's const-to-var switch required explicit
int32() casts at the few existing callers that pass these values to typed
BdevNvme* RPC wrappers.

+++ pkg/spdk/engine.go
@@ -2397,7 +2397,7 @@ func (e *Engine) BackupRestoreFinish(spdkClient *spdkclient.Client) error {
 		e.log.Infof("Attaching replica %s with address %s (transport=%s) before finishing restoration", replicaName, replicaAddress, transport)
 		nvmeBdevNameList, err := spdkClient.BdevNvmeAttachController(replicaName, helpertypes.GetNQN(replicaName), replicaIP, replicaPort,
 			transport.ToSPDKTransportType(), spdktypes.NvmeAddressFamilyIPv4,
-			int32(e.ctrlrLossTimeout), replicaReconnectDelaySec, int32(e.fastIOFailTimeoutSec), replicaMultipath)
+			int32(e.ctrlrLossTimeout), int32(replicaReconnectDelaySec), int32(e.fastIOFailTimeoutSec), replicaMultipath)
 		if err != nil {
 			return err
 		}
